WEB: Refactor create product page
